Executable Creation
The executable works! (sorta)

It opens up Bob, visuals (including avatars and styling) both appear to be working at this time.

However, ollama does not run at the start (yet). Overall, the executable that this makes *should* work, so long as the files it needs are someplace on the computer.